Simple app to demonstrate bug in Pixel launcher regarding defaultly disabled Activity aliases.
There are two Activities. MainActivity and SeparateActivity. SeparateActivity is not a launch activity. However it does have an Alias "AliasThang" which is a launch activity and is defaultly disabled.
If you click a button in MainActivity it will enable that alias.

If you create shortcuts on the launcher for both MainActivity and AliasThang and reboot the device, the shortcut that was initially pointing to AliasThang gets changed to target MainActivity.